Put implementation for platform-specific APIs into a common `env.cc`.
